machine: make machinectl list[-images] --full work as expected
Since systemd v239, both commands `machinectl list --full` and
`machinectl list-images --full` printed out a shorted name for each
machine name with an ellipsis character at the end. For example:

Before v239:
```
one-machine-with-a-long-name container systemd-nspawn - - 10.22.0.2...
```

Since v239:
```
one-machine-with-a-long-na… container systemd-nspawn - - 10.22.0.2…
```

Most users would have expected full machine names without ellipsis,
when a `--full` option is given.

To make those commands print out full machine names, we need to
make table-related functions take an option for full width. When the
`--full` option is given, we calculate the greatest possible width
for each column, to set the value to the display map of the table.
That way we can make the table print out in a wide format.